LPS-54317 Rename
