+++ portal-impl/test/integration/com/liferay/portal/kernel/portlet/bridges/mvc/bundle/actioncommand/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Description: Test Bundle for ActionCommandCacheTest
+Bundle-Description: Test Bundle for ActionCommandTest
 Bundle-Version: 1.0
 Private-Package: $(bundle.package).*
